refactor: separate PreparationTimeInputFieldList and PreparationTimeForm into different files
